pyproject.toml
[tool.poetry]
name = "oc_meta"
version = "2.1.3"
description = "OpenCitations Meta contains bibliographic metadata associated with the documents involved in the citations stored in the OpenCitations infrastructure. The OpenCitations Meta Software performs two main actions: a data curation of the provided CSV files and the generation of new RDF files compliant with the OpenCitations Data Model."
authors = ["Arcangelo Massari <arcangelomas@gmail.com>", "Fabio Mariani", "Simone Persiani <iosonopersia@gmail.com>"]
license = "ISC License"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
argparse = "^1.4.0"
lxml = ">=4.9.2,<6.0.0"
beautifulsoup4 = "^4.11.1"
python-dateutil = "^2.8.2"
PyYAML = "^6.0"
requests = "^2.27.1"
rdflib = "^7.4.0"
tqdm = "^4.64.0"
filelock = "^3.13"
psutil = "^7.0.0"
Pebble = "^5.0.3"
zstandard = "^0.21.0"
fakeredis = "^2.10.3"
ndjson = "^0.3.1"
oc-ds-converter = "^1.0.4"
ijson = "^3.2.3"
internetarchive = "^5.2.0"
zenodopy = "^0.3.0"
retrying = "^1.3.4"
orjson = "^3.10.7"
numpy = "^2.1.1"
pandas = "^2.2.3"
rdflib-ocdm = "^1.0.2"
time-agnostic-library = "^5.0.7"
virtuoso-utilities = "^1.5.3"
oc-ocdm = "^9.4.1"
sparqlite = "^1.0.0"

[tool.poetry.group.dev.dependencies]
wget = "^3.2"
requests-cache = "^0.9.4"
coverage = "^6.4.4"
coverage-badge = "^1.1.0"
autopep8 = "^2.0.0"
langdetect = "^1.0.9"
pycountry = "^22.3.5"
isort = "^6.0.1"
scipy = ">=1.10.0,<2.0.0"
matplotlib = "^3.10.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test = 'test.run_all_tests:main'